requirements.txt
# Bull-Bot requires Python 3.11+ (see .python-version).
# Schemas use PEP 604 union syntax (int | None) which Pydantic v2
# parses at class-creation time — 3.10+ is a hard minimum.
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.26.0
pyarrow>=14.0.0
scipy>=1.11.0
matplotlib>=3.8.0
tabulate>=0.9.0
pydantic>=2.6.0
anthropic>=0.34.0
pandas-market-calendars>=4.4.0
pytz>=2024.1
tenacity>=8.2.0
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-asyncio>=0.23.0
freezegun>=1.4.0
